further fix Directions + float Distance
Still not done with directions + Cleaned up a little bit of code.

+++ Map of Denmark/app/src/main/java/bfst22/vector/DijkstraSP.java
@@ -53,7 +53,7 @@ public class DijkstraSP {
 
     }
 
-    public double h(PolyPoint start, PolyPoint target){
+    public float h(PolyPoint start, PolyPoint target){
         Distance d = new Distance();
         return d.haversineFormula(start,target);
     }
